requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.7.2
nltk==3.8.1
rouge==1.0.1
requests==2.31.0
